initial commit
This isn't completed... It works, but it doesn't include the main
"feature" I want, linking to deleted posts. The problem is that I can't
figure out a good way to determine what page the post would be on, so I
can't actually create a link to it. I could probably create cookies, and
a reload a bunch of times... but I won't do that.... this is a long
description...